+++ core/src/main/groovy/nz/govt/natlib/tools/sip/generation/newspapers/NewspaperType.groovy
@@ -10,7 +10,6 @@ class NewspaperType {
     String PATH_TO_SPREADSHEET
     Map SUPPLEMENTS
     Map PARENT_SUPPLEMENTS
-    Map SUBSTITUTABLE_SUPPLEMENTS
     ArrayList IGNORE
     String REVISIONS
     boolean CASE_SENSITIVE
@@ -38,7 +37,6 @@ class NewspaperType {
         PATH_TO_SPREADSHEET = newspaperType["PATH_TO_SPREADSHEET"]
         SUPPLEMENTS = newspaperType["SUPPLEMENTS"] != null ? newspaperType["SUPPLEMENTS"] as Map : null
         PARENT_SUPPLEMENTS = newspaperType["PARENT_SUPPLEMENTS"] != null ? newspaperType["PARENT_SUPPLEMENTS"] as Map : null
-        SUBSTITUTABLE_SUPPLEMENTS = newspaperType["SUBSTITUTABLE_SUPPLEMENTS"] != null ? newspaperType["SUBSTITUTABLE_SUPPLEMENTS"] as Map : null
         IGNORE = newspaperType["IGNORE"] != null ? newspaperType["IGNORE"] as ArrayList : null
         REVISIONS = newspaperType["REVISIONS"]
         CASE_SENSITIVE = newspaperType["CASE_SENSITIVE"]